Refactor anagram solution
The algorithm is as follows:

1. Remove a word from the dictionary. This is the "base" word.
2. Loop through the remaining words in the dictionary. These are "pair"
   words.
3. If the "base" word plus a "pair" word is a valid anagram, add them to
   the list of anagram pairs.

Space efficiency:
ES6 introduced support for tail-call optimisation, so the use of a
recursive function should not be an issue. As such, the space efficiency
is `O(n)`, where `n` is the size of the dictionary.

Time efficiency:
Given a dictionary of size `n`, the time efficiency is:
`(n - 1) + (n - 2) ... (n - n)`

This may be written as `O(∑n)`.